# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Response": {
				"properties": {
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					},
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IReceipt": {
				"properties": {
					"receiptDate": {
						"type": "string"
					},
					"fromParty": {
						"type": "string"
					},
					"toParty": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"refNo": {
						"type": "string"
					},
					"isPayment": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"receiptDate",
					"fromParty",
					"toParty",
					"amount",
					"refNo",
					"isPayment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INotes": {
				"properties": {
					"fromParty": {
						"type": "string"
					},
					"toParty": {
						"type": "string"
					},
					"receiptDate": {
						"type": "string"
					},
					"products": {
						"items": {
							"properties": {
								"prices": {
									"type": "number",
									"format": "double"
								},
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"productId": {
									"type": "string"
								}
							},
							"required": [
								"prices",
								"quantity",
								"productId"
							],
							"type": "object"
						},
						"type": "array"
					},
					"shippingAddress": {
						"type": "string"
					},
					"isDeliveryNote": {
						"type": "boolean"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"fromParty",
					"toParty",
					"receiptDate",
					"products",
					"shippingAddress",
					"isDeliveryNote",
					"totalAmount"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "inventory-system",
		"version": "0.1.3",
		"description": "Inventory Management System",
		"license": {
			"name": "license"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/receipt/save": {
			"post": {
				"operationId": "Save",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Receipt"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReceipt"
							}
						}
					}
				}
			}
		},
		"/receipt/getAll": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Receipt"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/receipt/get": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Receipt"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/receipt/cancelReceipt": {
			"patch": {
				"operationId": "CancelReceipt",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Receipt"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"cancel": {
										"type": "boolean"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"cancel",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/receipt/delete": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Receipt"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notes/save": {
			"post": {
				"operationId": "Save",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Notes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INotes"
							}
						}
					}
				}
			}
		},
		"/notes/getAll": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Notes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notes/get": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Notes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notes/cancelNote": {
			"patch": {
				"operationId": "CancelNote",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Notes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"cancel": {
										"type": "boolean"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"cancel",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/notes/delete": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Notes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}